# src/controllers/comment.controller.ts
import { Get, Route, Tags, Post, Put, Delete, Body, Path } from "tsoa";
import { Comment } from "../models";
import {
  getComments,
  ICommentPayload,
  createComment,
  getComment,
  updateComment,
  deleteComment,
} from "../repositories/comment.repository";

@Route("comments")
@Tags("Comment")
export default class CommentController {
  @Get()
  public async getComments(): Promise<Array<Comment>> {
    return getComments();
  }

  @Post("/")
  public async createComment(@Body() body: ICommentPayload): Promise<Comment> {
    return createComment(body);
  }

  @Get("/:id")
  public async getComment(@Path() id: string): Promise<Comment | null> {
    return getComment(Number(id));
  }

  @Put("/:id")
  public async updateComment(@Path() id: string, @Body() body: ICommentPayload): Promise<Comment | null> {
    return updateComment(Number(id), body);
  }

  @Delete("/:id")
  public async deleteComment(@Path() id: string): Promise<Comment | null> {
    return deleteComment(Number(id));
  }
}